+++ sampleapp/src/main/java/com/ryanbrooks/expandablerecyclerviewsample/linear/horizontal/HorizontalParent.java
@@ -65,4 +65,9 @@ public class HorizontalParent implements ParentListItem, Serializable {
     public void setInitiallyExpanded(boolean initiallyExpanded) {
         mInitiallyExpanded = initiallyExpanded;
     }
+
+    @Override
+    public String toString() {
+        return mParentText;
+    }
 }